requirements.txt
s3prl
speechbrain
torch
torchaudio
scipy
numpy
sklearn
tqdm
pyyaml
librosa
fairseq
einops
fire
soundfile
